# Autenticacion/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: API de Autenticacion
  version: 1.0.0
  description: API para gestionar usuarios, incluyendo registro, login,cambio de contraseña y operaciones CRUD de autenticacion
paths:
...
  /api/usuarios/{codigo}:
    put:
      summary: Actualizar perfil de usuario
      description: Actualiza la información del perfil de un usuario existente
      tags:
        - Usuario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo
          required: true
          schema:
            type: string
          description: Código del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarUsuarioDTO'
      responses:
        '200':
          description: Usuario actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleUsuarioDTO'
        '400':
          description: Datos de actualización inválidos o incompletos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '409':
          description: Usuario ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'


    delete:
      summary: Eliminar cuenta de usuario
      description: Elimina la cuenta de un usuario del sistema
      tags:
        - Usuario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo
          required: true
          schema:
            type: string
          description: Código del usuario a eliminar
      responses:
        '200':
          description: Usuario eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '403':
          description: Usuario inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'

    get:
      summary: Obtener detalles de usuario
      description: Obtiene los detalles de un usuario específico
      tags:
        - Usuario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo
          required: true
          schema:
            type: string
          description: Código del usuario
      responses:
        '200':
          description: Detalles de usuario obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleUsuarioDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeDTO'
...
components:
  schemas:
    MensajeDTO:
      type: object
      properties:
        error:
          type: boolean
        respuesta:
          type: T
        required:
          - error
          - respuesta
...
    ActualizarUsuarioDTO:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    DetalleUsuarioDTO:
      type: object
      properties:
        codigo:
          type: string
        email:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT